Function rock_paper_scissors::hands::play_hand

pub fn play_hand(own_hand: Hand, other_hand: Hand) -> HandResult

Determines the result of two hands being played against each other.

own_hand - The hand being played.

other_hand - The opponent’s hand.

Returns the result of the hands being played, as a HandResult./// Plays a round of Scissors, Paper, Rock with the given hands and returns the result.